MA-19378 eliminate a build error, stop building when errors occur
Pass $(MAKE) and cp commands in vvcam recipe to shell individually, then
a single failure will cause the build process to stop, otherwise, the cp
command execution success may hide the failure of $(MAKE) command
execution failure.

Make the file $(VVCAM_KERNELENVSH) to be a phony target, then the file
can be updated everytime. this helps when the file exists but the export
commands being redirected to the file have been changed.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: I91db1714b019f6381ac58ead2029f754ef1941c5