LF-8977: Remove the native mode of the new tuningtool(tuningext).
Before the VSI P22 release, old version tuning tool(tuning server)
is supported by the native mode, but the new version of the tuning tool
(tuningext) from the VSI P22 release doesn't need the native mode.

So this patch is to remove the native mode and the no use code for the
tuningext.

Signed-off-by: Alice Yuan <alice.yuan@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>